Move endpoint base to the `entity` package
